Feature/domain crud

+++ src/main/java/nl/syntouch/dmn/studio/resource/DMNResource.java
@@ -7,5 +7,5 @@ import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import static nl.syntouch.dmn.studio.DmnStudioConstants.*;
 
 @RolesAllowed({ROLE_ADMIN, ROLE_DEVELOPER, ROLE_APPROVER, ROLE_DEPLOYER, ROLE_READ})
-public interface DMNResource extends PanacheEntityResource<DMN, Integer> {
+public interface DMNResource extends PanacheEntityResource<DMN, Long> {
 }
