+++ src/main/java/nl/syntouch/dmn/studio/resource/TestResource.java
@@ -7,5 +7,5 @@ import nl.syntouch.dmn.studio.model.Test;
 import static nl.syntouch.dmn.studio.DmnStudioConstants.*;
 
 @RolesAllowed({ROLE_ADMIN, ROLE_DEVELOPER, ROLE_DEPLOYER})
-public interface TestResource extends PanacheEntityResource<Test, Integer> {
+public interface TestResource extends PanacheEntityResource<Test, Long> {
 }
